This is a developer experience issue. It doesn't affect any existing features but lowers CI times significantly.

[TestSkiplistChaos](https://github.com/heroiclabs/nakama/blob/master/internal/skiplist/skiplist_chaos_test.go#L39) has always been using the [same seed](https://github.com/heroiclabs/nakama/blob/master/internal/skiplist/skiplist_chaos_test.go#L40) (see analysis below) since [it was introduced 4 months ago](https://github.com/heroiclabs/nakama/commit/c5102ec08c0c09ff5ae62dec20328877e7e250d5) and ran sequentially.

By making the individual iterations run in parallel a significant speedup can be obtained (test takes ~80% less time to run).

- Old version: **92.10s** (avg 12 runs, my local machine).
- New version: **16.76s** (avg 12 runs, my local machine).

On GitHub Actions machines this was taking 175 seconds (avg, see analysis below) it should be much lower now (extrapolating same boost, it should take now ~35 seconds).

## GitHub Actions Analysis (previous version)

https://github.com/heroiclabs/nakama/actions/runs/10116183915/job/27978481102
` ok  	github.com/heroiclabs/nakama/v3/internal/skiplist	175.273s` entire test suite took 325s
https://github.com/heroiclabs/nakama/actions/runs/10123008945/job/27995805457
` ok  	github.com/heroiclabs/nakama/v3/internal/skiplist	176.145s` entire test suite took 322s
https://github.com/heroiclabs/nakama/actions/runs/10123130877/job/27996059606
` ok  	github.com/heroiclabs/nakama/v3/internal/skiplist	173.110s` entire test suite took 317s

All runs report the same output:
```
=== RUN   TestSkiplistChaos
** iteration=0, users=1374, ops=35005, lists=38
** iteration=1, users=890, ops=32301, lists=27
** iteration=2, users=2049, ops=38384, lists=32
** iteration=3, users=1746, ops=60205, lists=46
** iteration=4, users=228, ops=19687, lists=9
** iteration=5, users=489, ops=52435, lists=40
** iteration=6, users=434, ops=92996, lists=28
** iteration=7, users=276, ops=19057, lists=37
** iteration=8, users=1676, ops=55060, lists=32
** iteration=9, users=1944, ops=39779, lists=20
** iteration=10, users=1157, ops=50303, lists=49
** iteration=11, users=1222, ops=63424, lists=6
** iteration=12, users=1519, ops=10031, lists=13
** iteration=13, users=1545, ops=98758, lists=35
** iteration=14, users=1362, ops=62395, lists=37
** iteration=15, users=1170, ops=83115, lists=38
** iteration=16, users=1789, ops=55313, lists=44
** iteration=17, users=1545, ops=3161, lists=43
** iteration=18, users=230, ops=44652, lists=11
** iteration=19, users=2021, ops=44110, lists=12
--- PASS: TestSkiplistChaos (174.09s)
```

## Rationale / Why bother?

This package is the one taking the most time to run tests for by a laaaaarge margin:

```
ok  	github.com/heroiclabs/nakama/v3/internal/skiplist	178.018s
ok  	github.com/heroiclabs/nakama/v3/internal/gopher-lua	 32.952s
ok  	github.com/heroiclabs/nakama/v3/server	                 16.676s
ok  	github.com/heroiclabs/nakama/v3/internal/cronexpr	  1.049s
ok  	github.com/heroiclabs/nakama/v3/internal/satori	          1.009s
```

**EDIT**: Updated execution times, I was running `go test` without `-race`. The performance boost is still 80%