@toqueteos 2024-07-30
No need to use `math/rand/v2`. Just by running the test in parallel nets the most bang for the buck without requiring changes in go.mod or updating the vendor folder. I'll update the PR.